Fixed the demo database by adding isDeleted column in the many-to-many relationship tables and added the proficiency column in the table ProficiencyCharacter
