+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/BaseEntity.java
@@ -7,24 +7,14 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 
 @MappedSuperclass
-public class BaseEntity{
+public class BaseEntity extends SoftDeletable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
     protected Long id;
-    @Column
-    protected boolean isDeleted;
 
     public Long getId() {
         return id;
-    }
-
-    public boolean isDeleted() {
-        return isDeleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        isDeleted = deleted;
     }
 
 }
