+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/DNDclass.java
@@ -2,7 +2,7 @@ package com.example.dndCharacterCreatorDemo.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToMany;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
 import java.util.Set;
@@ -16,8 +16,8 @@ public class DNDclass extends BaseEntity{
     private String description;
     @Column
     private int hitDice;
-    @ManyToMany(mappedBy = "dndClasses")
-    private Set<Proficiency> proficiencies;
-    @ManyToMany(mappedBy = "dndClasses")
-    private Set<Spell> spells;
+    @OneToMany(mappedBy = "dndClass")
+    private Set<ProficiencyClass> proficiencyClasses;
+    @OneToMany(mappedBy = "dndClass")
+    private Set<SpellClass> spellClasses;
 }
