+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/SoftDeletable.java
@@ -1,19 +1,18 @@
-//package com.example.dndCharacterCreatorDemo.entities;
-//
-//import jakarta.persistence.Column;
-//import jakarta.persistence.MappedSuperclass;
-//import org.springframework.stereotype.Component;
-//
-//@MappedSuperclass
-//public class SoftDeletable {
-//    @Column
-//    protected boolean isDeleted;
-//
-//    public boolean isDeleted() {
-//        return isDeleted;
-//    }
-//
-//    public void setDeleted(boolean deleted) {
-//        isDeleted = deleted;
-//    }
-//}
+package com.example.dndCharacterCreatorDemo.entities;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public class SoftDeletable {
+    @Column
+    protected boolean isDeleted;
+
+    public boolean isDeleted() {
+        return isDeleted;
+    }
+
+    public void setDeleted(boolean deleted) {
+        isDeleted = deleted;
+    }
+}
