+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/Spell.java
@@ -1,9 +1,8 @@
 package com.example.dndCharacterCreatorDemo.entities;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 
@@ -28,18 +27,20 @@ public class Spell extends BaseEntity{
     private int duration;
     @Column
     private String description;
-    @ManyToMany
-    @JoinTable(
-            name = "Spell_Characters",
-            joinColumns = { @JoinColumn(name = "spellId") },
-            inverseJoinColumns = { @JoinColumn(name = "characterId") }
-    )
-    private Set<Character> characters;
-    @ManyToMany
-    @JoinTable(
-            name = "Spell_Classes",
-            joinColumns = { @JoinColumn(name = "spellId") },
-            inverseJoinColumns = { @JoinColumn(name = "classId") }
-    )
-    private Set<DNDclass> dndClasses;
+//    @ManyToMany
+//    @JoinTable(
+//            name = "Spell_Characters",
+//            joinColumns = { @JoinColumn(name = "spellId") },
+//            inverseJoinColumns = { @JoinColumn(name = "characterId") }
+//    )
+    @OneToMany(mappedBy = "spell")
+    private Set<SpellCharacter> spellCharacters;
+//    @ManyToMany
+//    @JoinTable(
+//            name = "Spell_Classes",
+//            joinColumns = { @JoinColumn(name = "spellId") },
+//            inverseJoinColumns = { @JoinColumn(name = "classId") }
+//    )
+    @OneToMany(mappedBy = "spell")
+    private Set<SpellClass> spellClasses;
 }
