+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/SpellCharacter.java
@@ -1,21 +1,20 @@
 package com.example.dndCharacterCreatorDemo.entities;
 
+import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 
 @Entity
 @Table
-public class ProficiencyCharacter extends SoftDeletable {
+public class SpellCharacter extends SoftDeletable {
     @Id
     @ManyToOne
-    @JoinColumn(name="proficiencyId")
-    private Proficiency proficiency;
+    @JoinColumn(name="spellId")
+    private Spell spell;
     @Id
     @ManyToOne
     @JoinColumn(name="characterId")
     private Character character;
-    private boolean expertise;
 }
