+++ dndCharacterCreatorDemo/src/main/java/com/example/dndCharacterCreatorDemo/entities/SpellClass.java
@@ -1,18 +1,18 @@
 package com.example.dndCharacterCreatorDemo.entities;
 
+import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 
 @Entity
 @Table
-public class ProficiencyClass extends SoftDeletable {
+public class SpellClass extends SoftDeletable{
     @Id
     @ManyToOne
-    @JoinColumn(name="proficiencyId")
-    private Proficiency proficiency;
+    @JoinColumn(name="spellId")
+    private Spell spell;
     @Id
     @ManyToOne
     @JoinColumn(name="classId")
